entity_alignment/requirements.txt
py2neo==2024.1.0
sentence-transformers>=2.6.0
torch>=2.0.0,<2.2.0
scikit-learn>=1.3.0
openai>=1.12.0
python-dotenv>=1.0.0
tqdm>=4.66.1
huggingface-hub>=0.20.0
transformers>=4.38.0